Date: Wednesday Oct 24
    Making code easier to read
