--- CityWeather.py
@@ -1,5 +1,0 @@
-class CityWeather:
-    def __init__(self,temperature,humidity,pressure):
-        self.temperature = temperature
-        self.humidity = humidity
-        self.pressure = pressure
