+++ app.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from flask import Flask,jsonify
+from flask import Flask,json,jsonify
 from urllib.request import urlopen
 from bs4 import BeautifulSoup as soup
 import re
-import CityWeather
 app = Flask(__name__)
 WEATHER_URL = 'https://www.timeanddate.com/weather/'
 
@@ -16,8 +15,8 @@ def home():
 
 @app.route('/<string:country>/<string:city>')
 def get_weather(country, city):
-    report = get_report(country, city)
-    return jsonify(report)
+    weather = get_report(country, city)
+    return jsonify(weather.__dict__)
 
 
 def to_num(text):
@@ -27,10 +26,10 @@ def to_num(text):
 def scrape_info(for_soup):
 
     page_soup = soup(for_soup, "html.parser")
-    temperature = page_soup.find_all("div", {"id": "qlook"})[0].find_all('div')
+    temperature = to_num(page_soup.find_all("div", {"id": "qlook"})[0].find_all('div')[1].text)
     humid_press = page_soup.find_all("div", {"id": "qfacts"})[0].find_all('p')
-    humidity = humid_press[4]
-    pressure = humid_press[5]
+    pressure = to_num(humid_press[4].text)
+    humidity = to_num(humid_press[5].text)
     city_weather = CityWeather(temperature, humidity, pressure)
     return city_weather
 
@@ -43,7 +42,8 @@ def get_report(country,city):
     try:
         weather_url = get_weather_url(country, city)
         request_website = urlopen(weather_url)
-        report = scrape_info(request_website.read())
+        report = scrape_info(request_website)
+
     except:
         print("Error: City couldn't be found")
     return report
@@ -51,3 +51,11 @@ def get_report(country,city):
 
 if __name__ == '__main__':
     app.run()
+
+
+class CityWeather:
+    def __init__(self,temperature,humidity,pressure):
+        self.temperature = temperature
+        self.humidity = humidity
+        self.pressure = pressure
+
